requirements.txt
pylint
requests
django
numpy
pillow
piexif
h5py
